requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreML==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Vision==11.0
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
